Fix missing glyphs on the default style
The default style uses some glyphs for the config (ie battery full) that are
depending on the private use area of Unicode. The style defines only Source Sans
Pro as a font which does not contain these glyphs, and instead displays Chinese
glyphs.

This fixes part of this, but an SR making the waybar package depend on the
fontawesome-font package is also needed.